Tweak how we use spans (#42)
* Tweak how we use spans.
- Base implementations now have virtual methods that can be overridden as appropriate to provide optimized versions for any of the following layouts:
    - [x0, y0, z0, x1, y1, z1, ..., xn, yn, zn]
    - [x0, y0, x1, y1, ..., xn, yn, z0, z1, ..., zn]
    - [x0, x1, ..., xn, y0, y1, ..., yn, z0, z1, ..., zn]

All built-in algorithms now only override the one-by-one variants.  The point of allowing subclasses to override for spans (and allowing different layouts) is to allow implementations that use SIMD and/or auxiliary DLLs for their implementations.  There ought not to be a **particularly** significant advantage to using spans the way we were doing before.

* Bleh, this was supposed to use the arrays, not the spans, to be a tiny bit cheaper on the crusty ol' .NET Framework with basically no extra development effort